+++ back/index.js
@@ -7,6 +7,7 @@ const SteamGames = require('./models/steamGame');
 const getUpdateById = require('./routes/getUpdateById');
 const getGameByID = require('./routes/getGameById');
 const getGameByName = require('./routes/getGameByName');
+const postRecc = require('./routes/postRecc');
 
 const PORT = process.env.PORT || 3001;
 const MONGODB_URL = process.env.MONGO || 'mongodb://localhost:27017/gamr';
@@ -95,39 +96,7 @@ app.get('/api/status', (req, res) => {
 app.use(getUpdateById);
 app.use(getGameByID);
 app.use(getGameByName)
-
-app.post('/steam/recc', (req, res) => {
-  const { developer, publisher, tags, appid } = req.body;
-  let queryBase = {
-    'steam.type': 'game',
-    appid: { $not: { $eq: appid } },
-  };
-  if (developer.length > 0)
-    queryBase = Object.assign(queryBase, {
-      'steam.developers': { $in: developer },
-    });
-  if (publisher.length > 0)
-    queryBase = Object.assign(queryBase, {
-      'steam.publishers': { $in: publisher },
-    });
-  tags.forEach(tag => {
-    queryBase = Object.assign(queryBase, {
-      [`tags.${tag}`]: { $exists: true },
-    });
-  });
-  try {
-    SteamGames.find(queryBase)
-      .sort({ [`tags.${tags[0]}`]: -1 })
-      .limit(100)
-      .exec((err, games) => {
-        if (err) console.error(err);
-        res.send(games).status(200);
-      });
-  } catch (error) {
-    console.error(error);
-    res.status(500);
-  }
-});
+app.use(postRecc);
 
 app.listen(PORT, () => {
   console.log(`Server listening on ${PORT}`);
